gojo_telegram_api/requirements.txt
telethon
python_dotenv
urllib3==1.26.6 
openai
beautifulsoup4
python-socks==2.3.0
db-sqlite3
rapidfuzz
pygsheet